EP extended endoints + I/O DTOs

+++ src/main/java/com/github/CodeNekomancer/OADA_Backend/persistence/service/EPService.java
@@ -1,5 +1,8 @@
 package com.github.CodeNekomancer.OADA_Backend.persistence.service;
 
+import com.github.CodeNekomancer.OADA_Backend.configurations.ExceptionManager.NotFound.EPNotFoundException;
+import com.github.CodeNekomancer.OADA_Backend.configurations.ExceptionManager.NotFound.UAccNotFoundException;
+import com.github.CodeNekomancer.OADA_Backend.model.EP.DTOs.EPOutputDTO;
 import com.github.CodeNekomancer.OADA_Backend.model.EP.EP;
 import com.github.CodeNekomancer.OADA_Backend.model.UAcc.UAcc;
 import com.github.CodeNekomancer.OADA_Backend.persistence.repository.EPRepository;
@@ -12,7 +15,8 @@ public class EPService extends BaseService<EP, Long, EPRepository> {
     @Autowired
     private UAccService uaccSrvc;
 
-    public Boolean addEPSrvc(UAcc uAcc) {
+    public Boolean addSrvc(UAcc uAcc) {
+        if (uaccSrvc.findById(uAcc.getUacc_ID()).isEmpty()) throw new UAccNotFoundException();
         EP ep = new EP();
         ep.setItsUAcc(uAcc);
         ep.setName(uAcc.getItsUniverse().getServerId() + "_" + uAcc.getItsADAcc().getUsername());
@@ -20,8 +24,14 @@ public class EPService extends BaseService<EP, Long, EPRepository> {
         return this.repo.findById(ep.getEp_ID()).isPresent();
     }
 
-    public Boolean delEPSrvc(Long id) {
+    public Boolean delSrvc(Long id) {
+        if (this.repo.findById(id).isEmpty()) throw new EPNotFoundException();
         this.repo.deleteById(id);
         return !this.repo.existsById(id);
+    }
+
+    public EPOutputDTO getSngSrvc(Long id) {
+        if (this.repo.findById(id).isEmpty()) throw new EPNotFoundException();
+        return new EPOutputDTO(this.repo.findById(id).get());
     }
 }
